More Garmin updates

+++ src/net/sf/jdivelog/model/garmin/GarminDiveSettings.java
@@ -11,13 +11,13 @@ public class GarminDiveSettings {
    /**
     * Water density in kg/m³
     */
-   private Float m_water_density = 1000.0f;
+   private Double m_water_density = 1000.0;
 
-   public Float getM_water_density() {
+   public Double getM_water_density() {
       return m_water_density;
    }
 
-   public void setM_water_density(Float m_water_density) {
+   public void setM_water_density(Double m_water_density) {
       this.m_water_density = m_water_density;
    }
 }
